requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Jinja2==3.0.3
Werkzeug==2.0.2
PyMySQL==1.0.2